+++ mindmap/mindmap.py
@@ -2,19 +2,29 @@
 
 from __future__ import annotations
 
+import json
 import logging
 
 import pkg_resources
+from django.core.exceptions import PermissionDenied
 from django.utils import translation
 from web_fragments.fragment import Fragment
 from xblock.core import XBlock
-from xblock.fields import Boolean, Dict, Scope, String
+from xblock.exceptions import JsonHandlerError
+from xblock.fields import Boolean, DateTime, Dict, Float, Integer, Scope, String
 from xblockutils.resources import ResourceLoader
 
-from mindmap.utils import _
+from mindmap.edxapp_wrapper.student import user_by_anonymous_id
+from mindmap.edxapp_wrapper.xmodule import get_extended_due_date
+from mindmap.utils import _, utcnow
+
 
 log = logging.getLogger(__name__)
 loader = ResourceLoader(__name__)
+
+ITEM_TYPE = "mindmap"
+ATTR_KEY_ANONYMOUS_USER_ID = 'edx-platform.anonymous_user_id'
+ATTR_KEY_USER_ROLE = 'edx-platform.user_role'
 
 
 @XBlock.wants("user")
@@ -23,6 +33,10 @@ class MindMapXBlock(XBlock):
     """
     Mind Map XBlock provides a way to create and save mind maps in a course.
     """
+
+    has_score = True
+    icon_class = "problem"
+
     display_name = String(
         display_name=_("Display name"),
         default="Mind Map",
@@ -56,8 +70,9 @@ class MindMapXBlock(XBlock):
                 {
                     "id": "root",
                     "isroot": "true",
-                    "topic": "Root"}
-                ]
+                    "topic": "Root",
+                }
+            ]
         },
         scope=Scope.settings,
     )
@@ -71,6 +86,58 @@ class MindMapXBlock(XBlock):
         default={},
         scope=Scope.user_state,
     )
+
+    weight = Float(
+        display_name=_("Problem Weight"),
+        help=_(
+            "Defines the number of points each problem is worth. "
+            "If the value is not set, the problem is worth the sum of the "
+            "option point values."
+        ),
+        values={"min": 0, "step": 0.1},
+        scope=Scope.settings,
+    )
+
+    points = Integer(
+        display_name=_("Maximum score"),
+        help=_("Maximum grade score given to assignment by instructors."),
+        default=100,
+        scope=Scope.settings,
+    )
+
+    submitted = Boolean(
+        display_name=_("Submitted"),
+        help=_("Whether the student has submitted their submission."),
+        default=False,
+        scope=Scope.user_state,
+    )
+
+    @property
+    def block_id(self):
+        """
+        Return the usage_id of the block.
+        """
+        return str(self.scope_ids.usage_id)
+
+    @property
+    def block_course_id(self):
+        """
+        Return the course_id of the block.
+        """
+        return str(self.course_id)
+
+    @property
+    def score(self):
+        """
+        Return score from submissions.
+        """
+        return self.get_score()
+
+    def max_score(self):
+        """
+        Return the maximum score.
+        """
+        return self.points
 
     def get_current_user(self):
         """
@@ -134,6 +201,8 @@ class MindMapXBlock(XBlock):
             "xblock_id": self.scope_ids.usage_id.block_id,
             "is_static": self.is_static,
             "is_static_field": self.fields["is_static"],
+            "can_submit_assignment": self.submit_allowed(),
+            "score": self.score,
         }
 
     def get_js_context(self, user, context):
@@ -167,6 +236,9 @@ class MindMapXBlock(XBlock):
         user = self.get_current_user()
         context = self.get_context(user)
 
+        if self.show_staff_grading_interface():
+            context["is_instructor"] = True
+
         frag = self.load_fragment("mindmap", context)
 
         frag.add_javascript(self.resource_string("static/js/src/requiredModules.js"))
@@ -187,7 +259,11 @@ class MindMapXBlock(XBlock):
         user = self.get_current_user()
         context = self.get_context(user)
         context.update({
-            "editable": True
+            "editable": True,
+            "points": self.points,
+            "points_field": self.fields["points"],
+            "weight": self.weight,
+            "weight_field": self.fields["weight"],
         })
 
         frag = self.load_fragment("mindmap_edit", context)
@@ -219,9 +295,6 @@ class MindMapXBlock(XBlock):
         """
         Return the current mind map content.
 
-        Args:
-
-
         Returns:
             dict: The current mind map content.
             None: If the file does not exist.
@@ -230,26 +303,326 @@ class MindMapXBlock(XBlock):
             return self.mindmap_student_body
         return self.mindmap_body
 
-    @XBlock.json_handler
-    def upload_file(self, data, _suffix="") -> None:
+    def show_staff_grading_interface(self) -> bool:
         """
-        Uploads a mind map file to S3.
+        Return if current user is staff and not in studio.
 
-        Args:
-            data (dict): The necessary data to upload the file
-            _suffix (str, optional): Defaults to "".
+        Returns:
+            bool: True if current user is instructor and not in studio.
         """
-        self.mindmap_student_body = data.get("mind_map")
-        return {"success": True}
+        in_studio_preview = self.scope_ids.user_id is None
+        return self.is_instructor() and not in_studio_preview
+
+    def is_instructor(self) -> bool:
+        """
+        Check if user role is instructor.
+
+        Returns:
+            bool: True if user role is instructor.
+        """
+        return self.get_current_user().opt_attrs.get(ATTR_KEY_USER_ROLE) == "instructor"
 
     @XBlock.json_handler
-    def studio_submit(self, data, _suffix=""):
+    def studio_submit(self, data, _suffix="") -> None:
         """
         Called when submitting the form in Studio.
+
+        Args:
+            data (dict): The necessary configuration data.
+            _suffix (str, optional): Defaults to "".
         """
         self.display_name = data.get("display_name")
         self.is_static = data.get("is_static")
         self.mindmap_body = data.get("mind_map")
+
+        # We need to validate the points and weight fields ourselves because
+        # Studio doesn't do it for us.
+        points = data.get("points", self.points)
+        weight = data.get("weight", self.weight)
+        self.points, self.weight = self.validate_score(points, weight)
+
+    @XBlock.json_handler
+    def save_assignment(self, data, _suffix="") -> dict:
+        """
+        Save a mind map JSON structure into the block state for the user.
+
+        Args:
+            data (dict): The necessary data to upload the file
+            _suffix (str, optional): Defaults to "".
+
+        Returns:
+            dict: A dictionary containing the handler result.
+        """
+        self.mindmap_student_body = data.get("mind_map")
+        return {
+            "success": True,
+        }
+
+    @XBlock.json_handler
+    def submit_assignment(self, data, _suffix="") -> dict:
+        """
+        Submit a student's saved submission. This prevents further saves for the
+        given block, and makes the submission available to instructors for grading
+
+        Args:
+            request (Request): The request object
+            _suffix (str, optional): Defaults to "".
+
+        Returns:
+            dict: A dictionary containing the handler result.
+        """
+        # Lazy import: import here to avoid app not ready errors
+        from submissions.api import create_submission  # pylint: disable=import-outside-toplevel
+
+        require(self.submit_allowed())
+
+        self.mindmap_student_body = data.get("mind_map")
+        answer = {
+            "mindmap_student_body": json.dumps(self.mindmap_student_body),
+        }
+        student_item_dict = self.get_student_item_dict()
+        create_submission(student_item_dict, answer)
+
+        self.submitted = True
+
+        return {
+            "success": True,
+        }
+
+    @XBlock.json_handler
+    def get_instructor_grading_data(self, _, _suffix="") -> dict:
+        """Return student assignment information for display on the grading screen.
+
+        Args:
+            request (Request): The request object.
+            _suffix (str, optional): Defaults to "".
+
+        Returns:
+            dict: A dictionary containing student assignment information.
+        """
+        require(self.is_instructor())
+
+        def get_student_data() -> dict:
+            """
+            Returns a dict of student assignment information along with
+            annotated file name, student id and module id, this
+            information will be used on grading screen
+            """
+            # Lazy import: import here to avoid app not ready errors
+            from submissions.models import StudentItem as SubmissionsStudent  # pylint: disable=import-outside-toplevel
+
+            students = SubmissionsStudent.objects.filter(
+                course_id=self.course_id, item_id=self.block_id
+            )
+            for student in students:
+                submission = self.get_submission(student.student_id)
+                if not submission:
+                    continue
+                user = user_by_anonymous_id(student.student_id)
+                score = self.get_score(student.student_id)
+                yield {
+                    "student_id": student.student_id,
+                    "submission_id": submission["uuid"],
+                    "answer_body": submission["answer"],
+                    "username": user.username,
+                    "timestamp": submission["created_at"].strftime(
+                        DateTime.DATETIME_FORMAT
+                    ),
+                    "score": score,
+                    "submitted": self.submitted,
+                }
+
+        return {
+            "assignments": list(get_student_data()),
+            "max_score": self.max_score(),
+            "display_name": self.display_name,
+        }
+
+    @XBlock.json_handler
+    def enter_grade(self, data, _suffix="") -> dict:
+        """
+        Persist a score for a student given by instructors.
+
+        Args:
+            data (dict): The necessary data to enter a grade to a specific submission.
+            _suffix (str, optional): Defaults to "".
+
+        Returns:
+            dict: A dictionary containing the handler result.
+        """
+        # Lazy import: import here to avoid app not ready errors
+        from submissions.api import set_score  # pylint: disable=import-outside-toplevel
+
+        require(self.is_instructor())
+
+        score = int(data.get("grade"))
+        uuid = data.get("submission_id")
+        if not score or not uuid:
+            raise JsonHandlerError(400, "Missing required parameters")
+        if score > self.max_score():
+            raise JsonHandlerError(400, "Score cannot be greater than max score")
+
+        set_score(uuid, score, self.max_score())
+
+        return {
+            "success": True,
+        }
+
+    @XBlock.json_handler
+    def remove_grade(self, data, _suffix="") -> dict:
+        """
+        Persist a score for a student given by instructors.
+
+        Args:
+            data (dict): The necessary data to remove the grade from a specific submission.
+            _suffix (str, optional): Defaults to "".
+
+        Returns:
+            dict: A dictionary containing the handler result.
+        """
+        # Lazy import: import here to avoid app not ready errors
+        from submissions.api import reset_score  # pylint: disable=import-outside-toplevel
+
+        require(self.is_instructor())
+
+        student_id = data.get("student_id")
+        if not student_id:
+            raise JsonHandlerError(400, "Missing required parameters")
+
+        reset_score(student_id, self.block_course_id, self.block_id)
+
+        return {
+            "success": True,
+        }
+
+    @staticmethod
+    def validate_score(points: int, weight: int) -> None:
+        """
+        Validate a score.
+
+        Args:
+            score (int): The score to validate.
+            max_score (int): The maximum score.
+        """
+        try:
+            points = int(points)
+        except ValueError as exc:
+            raise JsonHandlerError(400, "Points must be an integer") from exc
+
+        if points < 0:
+            raise JsonHandlerError(400, "Points must be a positive integer")
+
+        if weight:
+            try:
+                weight = float(weight)
+            except ValueError as exc:
+                raise JsonHandlerError(400, "Weight must be a decimal number") from exc
+            if weight < 0:
+                raise JsonHandlerError(400, "Weight must be a positive decimal number")
+
+        return points, weight
+
+    def submit_allowed(self) -> bool:
+        """
+        Return whether student is allowed to submit an assignment.
+        A student is allowed to submit an assignment if:
+        - The due date has not passed for the assignment
+        - The student has not already submitted
+        - The student has not already received a score
+
+        Returns:
+            bool: True if student is allowed to submit an assignment.
+        """
+        return (
+            not self.past_due()
+            and self.score is None
+            and not self.submitted
+        )
+
+    def get_score(self, student_id=None) -> int:
+        """
+        Return student's current score.
+
+        Args:
+            student_id (str, optional): The student id to get the score for.
+
+        Returns:
+            int: The student's current score.
+        """
+        # Lazy import: import here to avoid app not ready errors
+        from submissions.api import get_score  # pylint: disable=import-outside-toplevel
+
+        score = get_score(self.get_student_item_dict(student_id))
+        if score:
+            return score["points_earned"]
+
+        return None
+
+    def get_submission(self, student_id=None) -> dict:
+        """
+        Get student's most recent submission.
+
+        Args:
+            student_id (str, optional): The student id to get the submission for.
+
+        Returns:
+            dict: The student's most recent submission.
+        """
+        # Lazy import: import here to avoid app not ready errors
+        from submissions.api import get_submissions  # pylint: disable=import-outside-toplevel
+
+        submissions = get_submissions(
+            self.get_student_item_dict(student_id)
+        )
+        if submissions:
+            return submissions[0]
+
+        return None
+
+    def get_student_item_dict(self, student_id=None) -> dict:
+        """
+        Returns dict required by the submissions app for creating and
+        retrieving submissions for a particular student.
+
+        Args:
+            student_id (str, optional): The student id to get the student item for.
+
+        Returns:
+            dict: The student item dict.
+        """
+        if not student_id:
+            student_id = self.get_current_user().opt_attrs.get(ATTR_KEY_ANONYMOUS_USER_ID)
+
+        return {
+            "student_id": student_id,
+            "course_id": self.block_course_id,
+            "item_id": self.block_id,
+            "item_type": ITEM_TYPE,
+        }
+
+    def past_due(self) -> bool:
+        """
+        Return whether due date has passed.
+
+        Returns:
+            bool: True if due date has passed.
+        """
+        due = get_extended_due_date(self)
+        try:
+            graceperiod = self.graceperiod
+        except AttributeError:
+            # graceperiod and due are defined in InheritanceMixin
+            # It's used automatically in edX but the unit tests will need to mock it out
+            graceperiod = None
+
+        if graceperiod is not None and due:
+            close_date = due + graceperiod
+        else:
+            close_date = due
+
+        if close_date is not None:
+            return utcnow() > close_date
+        return False
 
     # TO-DO: change this to create the scenarios you'd like to see in the
     # workbench while developing your XBlock.
@@ -292,3 +665,11 @@ class MindMapXBlock(XBlock):
         Dummy method to generate initial i18n
         """
         return translation.gettext_noop('Dummy')
+
+
+def require(assertion):
+    """
+    Raises PermissionDenied if assertion is not true.
+    """
+    if not assertion:
+        raise PermissionDenied