+++ mindmap/static/js/src/mindmap.js
@@ -1,6 +1,12 @@
 // TODO: add notifications
 function MindMapXBlock(runtime, element, context) {
 
+    const saveMindMapURL = runtime.handlerUrl(element, "save_assignment");
+    const submitMindMapURL = runtime.handlerUrl(element, "submit_assignment");
+    const getGradingDataURL = runtime.handlerUrl(element, "get_instructor_grading_data");
+    const enterGradeURL = runtime.handlerUrl(element, "enter_grade");
+    const removeGradeURL = runtime.handlerUrl(element, "remove_grade");
+
     function showMindMap(jsMind, context) {
         const mind = context.mind_map
 
@@ -22,23 +28,68 @@ function MindMapXBlock(runtime, element, context) {
         currentMindMap.show(mind);
 
         $(element).find(`#save_button_${context.xblock_id}`).click(function () {
-            saveMindMap(runtime, element, currentMindMap, "student");
+            handleMindMap(runtime, element, currentMindMap, saveMindMapURL);
+        });
+
+        $(element).find(`#submit_button_${context.xblock_id}`).click(function () {
+            handleMindMap(runtime, element, currentMindMap, submitMindMapURL);
+        });
+
+        $(element).find(`#get_submissions_button_${context.xblock_id}`).click(function () {
+            $.post(getGradingDataURL, JSON.stringify({}))
+            .done(function (response) {
+                console.log(response.assignments);
+            })
+            .fail(function () {
+                console.log("Error submitting mind map");
+            });
+        });
+
+        $(element).find(`#enter-grade-button_${context.xblock_id}`).click(function () {
+            const grade = $(element).find(`#grade-input_${context.xblock_id}`).val();
+            const submission_id = $(element).find(`#submission-id-input_${context.xblock_id}`).val();
+            const data = {
+                "grade": grade,
+                "submission_id": submission_id,
+            }
+            console.log(data);
+            $.post(enterGradeURL, JSON.stringify(data))
+            .done(function (response) {
+                console.log(response);
+            })
+            .fail(function (error) {
+                console.log(error);
+            });
+        });
+
+        $(element).find(`#remove-grade-button_${context.xblock_id}`).click(function () {
+            const student_id = $(element).find(`#student-id-input_${context.xblock_id}`).val();
+            const data = {
+                "student_id": student_id,
+            }
+            console.log(data);
+            $.post(removeGradeURL, JSON.stringify(data))
+            .done(function (response) {
+                console.log(response);
+            })
+            .fail(function (error) {
+                console.log(error);
+            });
         });
 
     };
 
-    function saveMindMap(runtime, element, mindMap, path_prefix) {
+    function handleMindMap(runtime, element, mindMap, handlerUrl) {
         const mindMapData = mindMap.get_data("node_array");
         const jsonMindMapData = mindMapData;
-        const handlerUrl = runtime.handlerUrl(element, "upload_file");
-        const data = { mind_map: jsonMindMapData, path_prefix: path_prefix };
+        const data = { mind_map: jsonMindMapData };
 
         $.post(handlerUrl, JSON.stringify(data))
             .done(function (response) {
                 window.location.reload(false);
             })
             .fail(function () {
-                console.log("Error saving mind map");
+                console.log("Error submitting mind map");
             });
     }